Make TWRP compile for arm64
-Remove dosfstools for arm64 until we can make it compile
-Fix TW_USE_TOOLBOX flag to work again
-Fix symlinking and handling of sh when using mksh
-Fix legacy properties to find futex_wake function
-Fix libcrecovery to not use bsd_signal anymore
-Fix rules for building with regards to libcrecovery
-Fix toolbox rules to compile tools that still work in 5.0
-Fix a few compile errors specific to arm64

Note: toolbox removed several tools in 5.0 including dd which we
need for TWRP backup and restore. You will need to copy in an
older toolbox for TWRP to work. At this point I do not know of a
busybox that will compile for arm64 so toolbox is required.

Testers report that TWRP does not boot on Nexus 9 and we fail to
get a shell for adb shell. At least it compiles without errors.

+++ legacy_property_service.c
@@ -25,6 +25,8 @@
 #include <limits.h>
 #include <errno.h>
 
+#include "../../bionic/libc/private/bionic_futex.h"
+
 #include <cutils/properties.h>
 
 #include "legacy_properties.h"
@@ -181,7 +183,7 @@ static int legacy_property_set(const char *name, const char *value)
         memcpy(pi->value, value, valuelen + 1);
 
         pa->toc[pa->count] =
-            (namelen << 24) | (((unsigned) pi) - ((unsigned) pa));
+            (namelen << 24) | (((unsigned long) pi) - ((unsigned long) pa));
 
         pa->count++;
         pa->serial++;
